requirements.txt
anyio==3.6.2
autopep8==2.0.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
dnspython==2.2.1
ecdsa==0.18.0
fastapi==0.85.2
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
mysqlclient==2.1.1
passlib==1.7.4
psycopg2==2.9.5
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
watchfiles==0.18.0
websockets==10.4
